Use Entity object config new card as default
The original card uses the string name for entities by default, which causes many issues
and requires handling two configs in parallel. This change uses the object configuration
as the default and simplifies the logic in other methods - strict object everywhere.

This is a first step for adding a visual editor.